requirements.txt
torchvision==0.19.1
torch==2.4.1
openai==1.12.0
tqdm
opencv-python
omegaconf
python-dotenv 
matplotlib
lightning_fabric
transformers==4.44.2
open-clip-torch==2.24.0
pandas
imageio
decord
einops
av
diffusers==0.30.2
pytorch_lightning
pycocoevalcap
accelerate
kornia
Pillow==9.4.0
numpy==1.23.1
nltk
backoff 
imageio-ffmpeg
configparser